prototype-0529/requirements.txt
streamlit ='1.5.0'
pymysql
pillow
langchain
openai
langchain_openai
langchain_community ='0.2.1'

